+++ src/sxapi/base.py
@@ -5,7 +5,6 @@ from enum import Enum
 
 import requests
 
-PUBLIC_API_V2_BASE_URL = "https://api.smaxtec.com/api/v2"
 INTEGRATION_API_V2_BASE_URL = "https://api.smaxtec.com/integration/v2"
 
 
@@ -91,20 +90,6 @@ class BaseAPI(object):
         url = self.to_url(path)
         r = self.session.delete(url, *args, **kwargs)
         return r.json()
-
-
-class PublicAPIV2(BaseAPI):
-    def __init__(self, base_url=None, email=None, password=None, api_token=None):
-        """Initialize a new public api client instance."""
-        base_url = base_url or PUBLIC_API_V2_BASE_URL
-        api_type = ApiTypes.PUBLIC
-        super().__init__(
-            base_url,
-            email=email,
-            password=password,
-            api_token=api_token,
-            api_type=api_type,
-        )
 
 
 class IntegrationAPIV2(BaseAPI):
